requirement.txt
ir-datasets
modin[dask]
pyarrow
psycopg2-binary
python-dotenv